feat: Replace existing data when preloading Fog Nodes in Vault
This commit updates the Rake task responsible for preloading Fog Nodes in Vault. It now calculates the hashcode of the code and clears the existing data at the 'fog-nodes-v1/' endpoint before writing the new data.

The hashcode for the code is now determined just once and applied to all Fog Nodes, as they are based on the same implementation.

# Additions:
- Calculation of the code hashcode
- Deletion of existing data at the 'fog-nodes-v1/' endpoint